+++ assets/javascripts/kitten/components/icons/cb-icon/index.js
@@ -1,10 +1,12 @@
 import React from 'react'
+import classNames from 'classnames'
 
-export const CbIcon = props => (
+export const CbIcon = ({className, ...props}) => (
   <svg
     xmlns="http://www.w3.org/2000/svg"
     viewBox="0 0 2051.48 845.02"
     width="36"
+    className={classNames('k-ColorSvg', className)}
     {...props}
   >
     <title>CB</title>