+++ CHANGELOG.md
@@ -5,6 +5,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [unreleased]
 
 - Feature: `ButtonGroup` : Add `orion` variant.
+- Fix: `Button`: Color icons will not be styled by Button SVG styles.
 
 ## [3.15.0] - 2021-05-28
 
@@ -26,7 +27,6 @@ Fixes:
 - `DashboardMenu`: Fix list items spacing.
 - `Separator`: Reduce specificity of the styles.
 - `HeaderNav`: Fix accessibility error.
->>>>>>> origin
 
 ## [3.14.0] - 2021-05-25